arreglo import

+++ app/src/main/java/com/hans/gesticar/ui/components/DropdownTextField.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.toDp
+
 
 @Composable
 fun DropdownTextField(
@@ -58,7 +58,7 @@ fun DropdownTextField(
             modifier = Modifier
                 .fillMaxWidth()
                 .onGloballyPositioned { coordinates ->
-                    menuWidth = with(density) { coordinates.size.width.toDp() }
+                    menuWidth = with(density) { coordinates.size.width.toFloat().toDp() }
                 }
                 .clickable(
                     interactionSource = interactionSource,
